+++ docs/src/cache/tutorial004.py
@@ -3,5 +3,6 @@ from corva import Api, Cache, ScheduledDataTimeEvent, scheduled
 
 @scheduled
 def scheduled_app(event: ScheduledDataTimeEvent, api: Api, cache: Cache):
-    cache.set(key='str', value='text')  # <.>
-    assert cache.get(key='str') == 'text'  # <.>
+    cache.set(key='int', value=str(0))  # <.>
+    assert cache.get('int') == '0'  # <.>
+    assert int(cache.get('int')) == 0  # <.>